requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.4
docx2pdf==0.1.8
lxml==5.3.0
pypandoc==1.15
python-docx==1.1.2
pywin32==308
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
